Introduce replacement Funder model
The new Funder model no longer needs to be an ActiveRecord model.
Instead, it is smart enough to reach into the Answer records to assemble
an instance that represents a Funder.

There might be models within the application that still need to be
an ActiveRecord instance, but Funder is fairly simple and does not need
that.