Migrate FinancialDisclosureTask to CustomCardTask
We purposely created a new migrator to move data from the legacy Task
to CustomCardTask.  FinancialDisclosure was an odd bird.  It had
answers stored on the Funder model (attributes), the
FinancialDisclosureTask itself, and associated Answers (records).

So, this migrator was very much different than the one used for other
simpler migrations.  It also is very unique to Funders/FinancialDisclosure
so much of it is not reusable.