Minimal FinancialDisclosureTask deprecation
We cannot fully remove the FinancialDisclosureTask since we need it
for migrations and to have a safety net in case we need to fall back
to the legacy task work.

This surgically removes some methods in common areas so that tests will
properly fail if the legacy code is still being referenced.