APERTA-11959 make dueDatime a real model in ember
Previously we were doing a lot of model-like things with dueDatetimes in ember,
but it wasn't formally declared as a model. This fixes that and gets rid of
some weird workarounds.

The motivation for doing this was to serialize scheduledEvents along with
dueDatetimes so that when a dueDatetime is updated on the server, it sends
back updated scheduledEvents.