APERTA-10460 clean up JournalTaskType for all journals
The previous version of the `CustomCard::FinancialDisclosureMigrator` only deleted
a single `JournalTaskType` rather than all of them. This migration will clean up
the data on the QA machines where that version has already run, and fixes the
`CustomCard::FinancialDisclosureMigrator` to use the same process as the
`CustomCard::Migrator`

Overriding the `self.create_journal_task_type?` in
`TahiStandardTasks::FinancialDisclosureTask` isn't actually needed (when we've
left the class around post-migration we haven't had this issue), but the
belt-and-suspenders approach won't actively hurt anything. That task file will
get deleted in the next release.