APERTA-10186 correct author validation for newly completed task
The nature of the `task` assocation on `Author` made it such that when the
AuthorsTask was completed in the `TasksController#update` action the validations
that were meant to fire on the task's `Authors` weren't running.

When the Author is validated through the AssocationValidator, the task is only
complete in memory, not yet persisted. The `task` method on author before these
changes looked up a new instance of the task from the database, meaning that
author.task.completed would be false even though it read as true in the
controller.

The main thing this PR does is add an extra attr_accessor to `Author`, which the
AssocationValidator sets in a lambda.  This flag forces the validation
regardless of the task's completed status.  It falls back to 'task.completed'
for use cases outside of the TasksController

I also moved the `task` relationship back to a `has_one` rather than looking it
up directly.

Finally, the `inverse_of` option doesn't do anything on `through` relationships
so I took that out as I went.