APERTA-10186 strip whitespace on some author fields
Uses a before_validation callback. Email is the most necessary, as it affects
the validation (the Devise email regex doesn't like leading or trailing space).
The name fields are likely to be concatted together, so they're trimmed mostly
for aesthetic reasons.  We can easily add more fields to the list if we want.