APERTA-10186 fix author creator spec
There's an association caching problem of sorts that only affects the test code
where the authors task and the author are made in the same test. You can also
see this in the `author_spec.rb` file but I've called it out here for a commit
message.

```ruby
    let(:authors_task) { FactoryGirl.create(:authors_task, completed: true) }
    let(:author) { FactoryGirl.create(:author, :contributions, paper: authors_task.paper).reload }
```

Without the `.reload` at the end of the author let block, `author.task` will be
nil. Alternatively calling `author.association(:task).reload` and then
`author.task` will also display the correct result.  This doesn't affect the
actual circumstances where it's used but if it makes us nervous we can dig a bit
into the association code to see why it's happening.