APERTA-8127 remove dropped element from DOM
We use $.sortable() for our underlying sort implementation in the
`{{sortable-list}}` component. Previous to this, when we drop an
task in the list we'd see both the dropped element AND a new element
that ember would create as a result of updating the phase information
for that task.  Our sortable implementation uses $().index() to
calculate the positions of the items, so we have to actually remove the
dropped element from the DOM otherwise subsequent moves will fail.