Remove vestigial task.descendants spec
This used to be a spec for task.all_task_types, which we've subsequently
removed from the codebase.  The test as it stands is testing a standard
`ActiveSupport::DescendantsTracker` method, which we don't need to do.
It looks like the `ActionDispatch::Reloader` calls were the thing that
caused some failures related to missing constants after removing the
rails engines.