APERTA-11685 load ORCID accounts with users
Currently a race condition exists for the orcidIdentifier presence
validation as it's invoked from  the `authors-task` component.
If an admin opens the authors task and tries to complete it without
opening the author forms, the task runs validations on the authors.
During the process of validation, the system will call
`author.get('user.orcidAccount.identifier')`, which returns a promise.
Unfortunately our validations only run synchronously right now, if we
have to fetch the orcid account from the server we interpret that as
"not being present" (sorta).

This is the low-impact solution for our problem.  We don't load users
very frequently, and including the orcid account in the user payload
doesn't kick off any further queries beyond that.  Alternatives would
include trying to fetch the users when we load the authors card, or at
some other point in the UI lifecycle before we press the button to
complete the task.