APERTA-12390 fix author orcid serialization
The OrcidAccountSerializer uses the :current_user scope, which the
controller normally provides when the serializer is invoked as part of a
responder, like `respond_with: user` or a `render json: user`, but in
the `author_list_item` method we use `as_json`, which doesn't set the
scope on the serializer, thus causing a 500 when we tried to use it.

Also added a test that will fail on master. The specific case of having
`orcid_connect_enabled` set to true wasn't relevant until the fixes in
c45489093f0b19605f2eb5a7beacdc7c3c09f023, and we missed adding the test case
there.