APERTA-12318 update states for all roles
The CardPermissions.update_stateful_permissions method will
short-circuit some of its operations if it's passed a permission.
In the case of the controller's #update action that's not what we want,
as that logic is responsible for assigning different permission states
for different roles.  This was causing side effects like some users not
being able to edit cards when the paper was in a submitted state, even
though it looked like we were giving them permission to do so.