APERTA-10841 avoid liquid error when invitee is not defined
Invitation records have a nil invitee association when there isn't
a user present for the invited email. Previously, the invatation templates were
rendering with a Liquid Error string because the user context was attempting
to call full_name on a nil object which it was proxying. I tried using
{% if invitee } in the template, but it would still evaluate to true when
the invitee was absent, presumably because the context is truthy even when
it's wrapping a nil object.

This change sets subcontexts to nil when the object they're proxying is nil
so we can more naturally say things like {% if invitee %} in templates and
have them behave in an expected way (be falsey when they're nil).