Fix flaky paper tracker spec
This one was tricky. The test is creating a paper, then ensuring that
the query param is parsed as a DOI and the otherwise matching paper is
NOT returned.

Really, this is a bad test, because it mixes up two bits into one
test: the paper tracker query parser and the controller.

That said, the problem is that after a certain number of journals were
seeded, we ended up with an accidentally matching DOI (because of the
prefix).

The fix to this was to ensure that journal/publisher prefixes did not
have 123 in them.